api/requirements.txt
strawberry-graphql>=0.219.0
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
openai>=1.10.0
